requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-RBAC==0.5.0
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
packaging==21.3
pluggy==1.0.0
protobuf==4.21.2
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
result==0.8.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.0
